device info

+++ src/main/java/com/sundirect/crm/apirepo/SubscriberRepo.java
@@ -8,7 +8,7 @@ import com.sundirect.crm.apientity.MyplexUserUser;
 
 public interface SubscriberRepo extends JpaRepository<MyplexUserUser, Integer>{
 	
-	Optional<MyplexUserUser> findById(Long id);
+	Optional<MyplexUserUser> findById(Integer id);
 	
 	Optional<MyplexUserUser> findByMobileNo(Long mobileNo);
